feat: add ws integration right panel issues
